Refactor: Extract Parser class
