+++ data/database.txt
@@ -1,1 +1,4 @@
-D,0,finish this assignment ,01-01-2021 12:01
+T,1,to this as well
+T,1,and this
+D,1,this assignment ,01-01-2021 10:22
+T,0,show this in db
