+++ src/main/java/Parser.java
@@ -1,0 +1,103 @@
+import java.io.IOException;
+import java.util.Arrays;
+import java.util.LinkedList;
+import java.util.List;
+
+public class Parser {
+
+    /**
+     * Performs command based on String user input
+     * @param userInput string command for chatbot
+     * @throws InvalidArgumentException Invalid or no arguments given
+     * @throws TooManyArgumentsException Too many /by or /at connectors
+     */
+    public void parseCommand(String userInput, TaskList tasks, Ui ui, Storage storage ) throws InvalidArgumentException,
+            TooManyArgumentsException {
+        List<String> userInputList = new LinkedList<>(Arrays.asList(userInput.split(" ")));
+        BotCommand command = BotCommand.valueOf(userInputList.get(0).toUpperCase());
+        String description;
+        String[] userInputArgs;
+
+        switch (command) {
+        case BYE:
+            ui.showLine();
+            ui.sayBye();
+            ui.showLine();
+            break;
+        case LIST:
+            ui.showLine();
+            tasks.printList();
+            ui.showLine();
+            break;
+        case DONE:
+            ui.showLine();
+            tasks.markAsDone(Integer.parseInt(userInputList.get(1)));
+            ui.showLine();
+            break;
+        case DELETE:
+            // check delete argument
+            if (!isNumeric(userInputList.get(1))) {
+                throw new InvalidArgumentException("Delete argument is not numeric");
+            }
+            ui.showLine();
+            tasks.deleteTask(Integer.parseInt(userInputList.get(1)));
+            ui.showLine();
+            break;
+        case TODO:
+            userInputList.remove(0);
+            if (userInputList.size() == 0) {
+                throw new InvalidArgumentException("No arguments submitted for todo");
+            }
+            description = String.join(" ", userInputList);
+            ui.showLine();
+            tasks.createToDo(description);
+            ui.showLine();
+            break;
+        case DEADLINE:
+            userInputList.remove(0);
+            if (userInputList.size() == 0) {
+                throw new InvalidArgumentException("No arguments submitted for deadline");
+            }
+            //split description and by
+            userInputArgs = String.join(" ", userInputList).split("/by ");
+            if (userInputArgs.length > 2) {
+                throw new TooManyArgumentsException("Too many arguments given for deadline");
+            } else if (userInputArgs.length == 1) {
+                throw new InvalidArgumentException("Could not find connector /by ");
+            }
+            description = userInputArgs[0];
+            String by = userInputArgs[1];
+            ui.showLine();
+            tasks.createDeadline(description, by);
+            ui.showLine();
+            break;
+        case EVENT:
+            userInputList.remove(0);
+            if (userInputList.size() == 0) {
+                throw new InvalidArgumentException("No arguments submitted for event");
+            }
+            //split description and at
+            userInputArgs = String.join(" ", userInputList).split("/at ");
+            if (userInputArgs.length > 2) {
+                throw new TooManyArgumentsException("Too many arguments given");
+            } else if (userInputArgs.length == 1) {
+                throw new InvalidArgumentException("Could not find connector /at");
+            }
+            description = userInputArgs[0];
+            String at = userInputArgs[1];
+            ui.showLine();
+            tasks.createEvent(description, at);
+            ui.showLine();
+            break;
+        }
+    }
+
+    /**
+     * Helper function to check if string is numeric
+     * @param str string to test if numeric
+     * @return boolean
+     */
+    public static boolean isNumeric(String str) {
+        return str.matches("-?\\d+(\\.\\d+)?");
+    }
+}
